requirements.txt
einops>=0.3.2
matplotlib>=3.3.4
nibabel>=3.2.1
numpy>=1.21.0
pandas>=1.2.4
pillow>=8.2.0
python>=3.7.10
pytorch>=1.10.0
scikit-image>=0.18.1